move existing config token to a new config dir
