+++ src/app/guards/home-redirect.guard.ts
@@ -1,6 +1,6 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, UrlTree } from '@angular/router';
-import { DEFAULT_ACTIVITY_ROUTE } from '../models/routing/default-route-tokens';
+import { DEFAULT_ACTIVITY_ROUTE } from 'src/app/config/default-route-tokens';
 import { ItemRouter } from '../models/routing/item-router';
 
 export const homeRedirectGuard: CanActivateFn = (): UrlTree => {
